cleanup inspired by GCC build

+++ src/game/Object/ObjectMgr.cpp
@@ -8107,7 +8107,7 @@ void ObjectMgr::LoadActiveEntities(Map* _map)
     if (!_map)
     {
         uint32 continents[] = {0, 1};
-        for (int i = 0; i < countof(continents); ++i)
+        for (uint8 i = 0; i < countof(continents); ++i)
         {
             _map = sMapMgr.FindMap(continents[i]);
             if (!_map)
